pyproject.toml
[project]
name = "ai-diagram-service"
version = "0.1.0"
description = "AI-powered diagram generation service using LLM agents"
authors = [
    {name = "AI Engineer", email = "ai@example.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "langchain>=0.1.0",
    "langchain-google-genai>=0.0.5",
    "google-generativeai>=0.3.0",
    "diagrams>=0.23.3",
    "pillow>=10.0.0",
    "aiofiles>=23.2.0",
    "httpx>=0.25.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.12.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true 